+++ src/ctrl.rs
@@ -131,7 +131,7 @@ where
 ///
 /// This is called by UblkCtrlInner::new()/new_async() when the ring hasn't been
 /// initialized yet. Uses default values similar to the original approach.
-fn init_ctrl_task_ring_default(depth: u32) -> Result<(), UblkError> {
+pub(crate) fn init_ctrl_task_ring_default(depth: u32) -> Result<(), UblkError> {
     ublk_init_ctrl_task_ring(|ring_opt| {
         if ring_opt.is_none() {
             let ring = IoUring::<squeue::Entry128>::builder()
@@ -412,7 +412,7 @@ struct QueueAffinityJson {
 /// operations for ublk devices. This includes device information, queue data,
 /// target configuration, and runtime state management.
 #[derive(Debug)]
-struct UblkJsonManager {
+pub(crate) struct UblkJsonManager {
     /// Device ID for JSON file naming
     dev_id: u32,
     /// Current JSON data
@@ -714,8 +714,8 @@ impl UblkCtrlBuilder<'_> {
             self.dev_flags,
         )
     }
-    pub async fn build_async(self) -> Result<UblkCtrl, UblkError> {
-        UblkCtrl::new_async(
+    pub async fn build_async(self) -> Result<super::ctrl_async::UblkCtrlAsync, UblkError> {
+        super::ctrl_async::UblkCtrlAsync::new_async(
             Some(self.name.to_string()),
             self.id,
             self.nr_queues.into(),
@@ -743,25 +743,25 @@ pub struct UblkCtrl {
     inner: RwLock<UblkCtrlInner>,
 }
 
-struct UblkCtrlInner {
-    name: Option<String>,
+pub(crate) struct UblkCtrlInner {
+    pub(crate) name: Option<String>,
     file: fs::File,
-    dev_info: sys::ublksrv_ctrl_dev_info,
-    json_manager: UblkJsonManager,
-    features: Option<u64>,
+    pub(crate) dev_info: sys::ublksrv_ctrl_dev_info,
+    pub(crate) json_manager: UblkJsonManager,
+    pub(crate) features: Option<u64>,
 
     /// global flags, shared with UblkDev and UblkQueue
-    dev_flags: UblkFlags,
+    pub(crate) dev_flags: UblkFlags,
     // only set when running start_dev_async() which is enabled before
     // adding UBLK_CTRL_ASYNC_AWAIT
-    force_async: bool,
+    pub(crate) force_async: bool,
 
     //only used in Drop()
     force_sync: bool,
     cmd_token: i32,
     queue_tids: Vec<i32>,
     queue_selected_cpus: Vec<usize>,
-    nr_queues_configured: u16,
+    pub(crate) nr_queues_configured: u16,
 }
 
 /// Affinity management helpers
@@ -882,7 +882,7 @@ impl UblkCtrlInner {
     }
 
     /// Create appropriate affinity for queue thread setup
-    async fn create_thread_affinity_async(
+    pub(crate) async fn create_thread_affinity_async(
         &mut self,
         qid: u16,
     ) -> Result<UblkQueueAffinity, UblkError> {
@@ -1127,7 +1127,7 @@ impl UblkCtrlInner {
 
     /// Async legacy constructor wrapper for backward compatibility
     #[allow(clippy::too_many_arguments)]
-    async fn new_with_params_async(
+    pub(crate) async fn new_with_params_async(
         name: Option<String>,
         id: i32,
         nr_queues: u32,
@@ -1154,11 +1154,11 @@ impl UblkCtrlInner {
         (self.dev_info.flags & (super::sys::UBLK_F_UNPRIVILEGED_DEV as u64)) != 0
     }
 
-    fn get_cdev_path(&self) -> String {
+    pub(crate) fn get_cdev_path(&self) -> String {
         format!("{}{}", UblkCtrl::CDEV_PATH, self.dev_info.dev_id)
     }
 
-    fn for_add_dev(&self) -> bool {
+    pub(crate) fn for_add_dev(&self) -> bool {
         self.dev_flags.intersects(UblkFlags::UBLK_DEV_F_ADD_DEV)
     }
 
@@ -1175,17 +1175,17 @@ impl UblkCtrlInner {
         }
     }
 
-    fn store_queue_tid(&mut self, qid: u16, tid: i32) {
+    pub(crate) fn store_queue_tid(&mut self, qid: u16, tid: i32) {
         self.queue_tids[qid as usize] = tid;
     }
 
-    fn dump_from_json(&self) {
+    pub(crate) fn dump_from_json(&self) {
         self.json_manager.dump_json();
     }
 
     /// Returned path of this device's exported json file
     ///
-    fn run_path(&self) -> String {
+    pub(crate) fn run_path(&self) -> String {
         self.json_manager.get_json_path()
     }
 
@@ -1437,7 +1437,7 @@ impl UblkCtrlInner {
     /// logic as the synchronous version, choosing between DEL_DEV_ASYNC
     /// and DEL_DEV commands based on device flags.
     ///
-    async fn del_async_await(&mut self) -> Result<i32, UblkError> {
+    pub(crate) async fn del_async_await(&mut self) -> Result<i32, UblkError> {
         if self.is_deleted() {
             return Ok(0);
         }
@@ -1498,7 +1498,7 @@ impl UblkCtrlInner {
 
     /// Async version of read_dev_info() - retrieve device info from ublk driver
     ///
-    async fn read_dev_info_async(&mut self) -> Result<i32, UblkError> {
+    pub(crate) async fn read_dev_info_async(&mut self) -> Result<i32, UblkError> {
         match self.__read_dev_info2_async().await {
             Ok(result) => Ok(result),
             Err(_) => self.__read_dev_info_async().await,
@@ -1548,7 +1548,7 @@ impl UblkCtrlInner {
 
     /// Stop this device by sending command to ublk driver asynchronously
     ///
-    async fn stop_async(&mut self) -> Result<i32, UblkError> {
+    pub(crate) async fn stop_async(&mut self) -> Result<i32, UblkError> {
         let data = Self::prepare_stop_cmd();
         self.ublk_ctrl_cmd_async(&data).await
     }
@@ -1575,7 +1575,7 @@ impl UblkCtrlInner {
 
     /// Retrieve this device's parameter from ublk driver by
     /// sending command in async/.await
-    async fn get_params_async(&mut self, params: &mut sys::ublk_params) -> Result<i32, UblkError> {
+    pub(crate) async fn get_params_async(&mut self, params: &mut sys::ublk_params) -> Result<i32, UblkError> {
         let data = Self::prepare_get_params_cmd(params);
         self.ublk_ctrl_cmd_async(&data).await
     }
@@ -1602,7 +1602,7 @@ impl UblkCtrlInner {
     ///
     /// Note: device parameter has to send to driver before starting
     /// this device
-    async fn set_params_async(&mut self, params: &sys::ublk_params) -> Result<i32, UblkError> {
+    pub(crate) async fn set_params_async(&mut self, params: &sys::ublk_params) -> Result<i32, UblkError> {
         let mut p = *params;
 
         p.len = core::mem::size_of::<sys::ublk_params>() as u32;
@@ -1633,7 +1633,7 @@ impl UblkCtrlInner {
 
     /// Retrieving the specified queue's affinity from ublk driver in async/.await
     ///
-    async fn get_queue_affinity_async(
+    pub(crate) async fn get_queue_affinity_async(
         &mut self,
         q: u32,
         bm: &mut UblkQueueAffinity,
@@ -1652,7 +1652,7 @@ impl UblkCtrlInner {
         self.ublk_ctrl_cmd(&data)
     }
 
-    async fn __start_user_recover_async(&mut self) -> Result<i32, UblkError> {
+    pub(crate) async fn __start_user_recover_async(&mut self) -> Result<i32, UblkError> {
         let data = Self::prepare_start_user_recover_cmd();
         self.ublk_ctrl_cmd_async(&data).await
     }
@@ -1713,7 +1713,7 @@ impl UblkCtrlInner {
         Ok(0)
     }
 
-    async fn start_dev_async(&mut self, dev: &UblkDev) -> Result<i32, UblkError> {
+    pub(crate) async fn start_dev_async(&mut self, dev: &UblkDev) -> Result<i32, UblkError> {
         self.prep_start_dev_async(dev).await?;
 
         if self.dev_info.state != sys::UBLK_S_DEV_QUIESCED as u16 {
@@ -1841,7 +1841,7 @@ impl UblkCtrlInner {
         Ok(0)
     }
 
-    async fn build_json_async(&mut self, dev: &UblkDev) -> Result<i32, UblkError> {
+    pub(crate) async fn build_json_async(&mut self, dev: &UblkDev) -> Result<i32, UblkError> {
         // Update queue thread IDs if JSON already exists
         if self.update_json_queue_tids(dev) {
             return Ok(0);
@@ -1863,7 +1863,7 @@ impl UblkCtrl {
     /// char device and block device name may change according to system policy,
     /// such udev may rename it in its own namespaces.
     const CDEV_PATH: &'static str = "/dev/ublkc";
-    const BDEV_PATH: &'static str = "/dev/ublkb";
+    pub(crate) const BDEV_PATH: &'static str = "/dev/ublkb";
 
     const UBLK_DRV_F_ALL: u64 = (sys::UBLK_F_SUPPORT_ZERO_COPY
         | sys::UBLK_F_URING_CMD_COMP_IN_TASK
@@ -1959,7 +1959,7 @@ impl UblkCtrl {
     ///
     #[allow(clippy::too_many_arguments)]
     /// Validate parameters for new UblkCtrl creation
-    fn validate_new_params(
+    pub(crate) fn validate_new_params(
         flags: u64,
         dev_flags: UblkFlags,
         id: i32,
@@ -2029,69 +2029,7 @@ impl UblkCtrl {
         Self::new(None, id, 0, 0, 0, 0, 0, UblkFlags::empty())
     }
 
-    /// Async version of new() - creates a new ublk control device asynchronously
-    ///
-    /// # Arguments:
-    ///
-    /// * `name`: optional device name
-    /// * `id`: device id, or let driver allocate one if -1 is passed
-    /// * `nr_queues`: how many hw queues allocated for this device
-    /// * `depth`: each hw queue's depth
-    /// * `io_buf_bytes`: max buf size for each IO
-    /// * `flags`: flags for setting ublk device
-    /// * `tgt_flags`: target-specific flags
-    /// * `dev_flags`: global flags as userspace side feature
-    ///
-    /// This method performs the same functionality as new() but returns a Future
-    /// that resolves to the UblkCtrl instance. Most of the constructor work is
-    /// synchronous, so this mainly provides async compatibility.
-    ///
-    #[allow(clippy::too_many_arguments)]
-    pub async fn new_async(
-        name: Option<String>,
-        id: i32,
-        nr_queues: u32,
-        depth: u32,
-        io_buf_bytes: u32,
-        flags: u64,
-        tgt_flags: u64,
-        dev_flags: UblkFlags,
-    ) -> Result<UblkCtrl, UblkError> {
-        Self::validate_new_params(flags, dev_flags, id, nr_queues, depth, io_buf_bytes)?;
 
-        let inner = RwLock::new(
-            UblkCtrlInner::new_with_params_async(
-                name,
-                id,
-                nr_queues,
-                depth,
-                io_buf_bytes,
-                flags,
-                tgt_flags,
-                dev_flags,
-            )
-            .await?,
-        );
-
-        Ok(UblkCtrl { inner })
-    }
-
-    /// Async version of new_simple() - creates a simple UblkCtrl device asynchronously
-    ///
-    /// # Arguments:
-    ///
-    /// * `id`: device id (must be >= 0)
-    ///
-    /// This method performs the same functionality as new_simple() but returns a Future
-    /// that resolves to the UblkCtrl instance. The device can be used for deleting,
-    /// listing, recovering, etc., but not for adding new devices.
-    ///
-    pub async fn new_simple_async(id: i32) -> Result<UblkCtrl, UblkError> {
-        assert!(id >= 0);
-        // Simple constructor work is synchronous, so we just call the sync version
-        // This provides async compatibility for consistent API usage
-        Self::new_async(None, id, 0, 0, 0, 0, 0, UblkFlags::empty()).await
-    }
 
     /// Return current device info
     pub fn dev_info(&self) -> sys::ublksrv_ctrl_dev_info {
@@ -2180,31 +2118,13 @@ impl UblkCtrl {
         Ok(0)
     }
 
-    pub async fn configure_queue_async(
-        &self,
-        dev: &UblkDev,
-        qid: u16,
-        tid: i32,
-    ) -> Result<i32, UblkError> {
-        let mut ctrl = self.get_inner_mut();
-
-        ctrl.store_queue_tid(qid, tid);
-
-        ctrl.nr_queues_configured += 1;
-
-        if ctrl.nr_queues_configured == ctrl.dev_info.nr_hw_queues {
-            ctrl.build_json_async(dev).await?;
-        }
-
-        Ok(0)
-    }
 
     /// Dump this device info
     ///
     /// The 1st part is from UblkCtrl.dev_info, and the 2nd part is
     /// retrieved from device's exported json file
     /// Print device info after data has been collected
-    fn dump_device_info(ctrl: &UblkCtrlInner, p: &sys::ublk_params) {
+    pub(crate) fn dump_device_info(ctrl: &UblkCtrlInner, p: &sys::ublk_params) {
         let info = &ctrl.dev_info;
         println!(
             "\ndev id {}: nr_hw_queues {} queue_depth {} block size {} dev_capacity {}",
@@ -2253,37 +2173,6 @@ impl UblkCtrl {
         println!("\tublksrv_flags: 0x{:x}", ctrl.dev_info.ublksrv_flags);
     }
 
-    /// Dump this device info asynchronously
-    ///
-    /// This is the async version of dump(). The 1st part is from UblkCtrl.dev_info,
-    /// and the 2nd part is retrieved from device's exported json file.
-    /// Uses async I/O for driver communication and file operations.
-    pub async fn dump_async(&self) -> Result<(), UblkError> {
-        let mut ctrl = self.get_inner_mut();
-        let mut p = sys::ublk_params {
-            ..Default::default()
-        };
-
-        ctrl.read_dev_info_async().await.map_err(|e| {
-            error!(
-                "Dump dev {} failed: read_dev_info_async\n",
-                ctrl.dev_info.dev_id
-            );
-            e
-        })?;
-
-        ctrl.get_params_async(&mut p).await.map_err(|e| {
-            error!(
-                "Dump dev {} failed: get_params_async\n",
-                ctrl.dev_info.dev_id
-            );
-            e
-        })?;
-
-        Self::dump_device_info(&ctrl, &p);
-        ctrl.dump_from_json();
-        Ok(())
-    }
 
     pub fn run_dir() -> String {
         String::from("/run/ublksrvd")
@@ -2311,15 +2200,6 @@ impl UblkCtrl {
         self.get_inner_mut().read_dev_info()
     }
 
-    /// Retrieving device info from ublk driver in async/.await
-    ///
-    /// This method performs the same functionality as read_dev_info() but returns a Future
-    /// that resolves to the result. It uses the same fallback mechanism as the synchronous
-    /// version, trying UBLK_U_CMD_GET_DEV_INFO2 first and falling back to UBLK_U_CMD_GET_DEV_INFO.
-    ///
-    pub async fn read_dev_info_async(&self) -> Result<i32, UblkError> {
-        self.get_inner_mut().read_dev_info_async().await
-    }
 
     /// Retrieve this device's parameter from ublk driver by
     /// sending command
@@ -2329,17 +2209,6 @@ impl UblkCtrl {
         self.get_inner_mut().get_params(params)
     }
 
-    /// Retrieve this device's parameter from ublk driver by
-    /// sending command in async/.await
-    ///
-    /// This method performs the same functionality as get_params() but returns a Future
-    /// that resolves to the result. It uses the async uring infrastructure to avoid
-    /// blocking the calling thread while waiting for the ublk driver response.
-    ///
-    /// Can't pass params by reference(&mut), why?
-    pub async fn get_params_async(&self, params: &mut sys::ublk_params) -> Result<i32, UblkError> {
-        self.get_inner_mut().get_params_async(params).await
-    }
 
     /// Send this device's parameter to ublk driver
     ///
@@ -2349,16 +2218,6 @@ impl UblkCtrl {
         self.get_inner_mut().set_params(params)
     }
 
-    /// Send this device's parameter to ublk driver asynchronously
-    ///
-    /// This method performs the same functionality as set_params() but returns a Future
-    /// that resolves to the result. It uses the async uring infrastructure to avoid
-    /// blocking the calling thread while waiting for the ublk driver response.
-    ///
-    /// Note: device parameter has to send to driver before starting this device
-    pub async fn set_params_async(&self, params: &sys::ublk_params) -> Result<i32, UblkError> {
-        self.get_inner_mut().set_params_async(params).await
-    }
 
     /// Retrieving the specified queue's affinity from ublk driver
     ///
@@ -2366,23 +2225,6 @@ impl UblkCtrl {
         self.get_inner_mut().get_queue_affinity(q, bm)
     }
 
-    /// Retrieving the specified queue's affinity from ublk driver in async/.await
-    ///
-    /// This method performs the same functionality as get_queue_affinity() but returns a Future
-    /// that resolves to the result. It uses the async uring infrastructure to avoid
-    /// blocking the calling thread while waiting for the ublk driver response.
-    ///
-    /// # Arguments
-    /// * `q` - Queue ID
-    /// * `bm` - UblkQueueAffinity to populate with the affinity bitmap
-    ///
-    pub async fn get_queue_affinity_async(
-        &self,
-        q: u32,
-        bm: &mut UblkQueueAffinity,
-    ) -> Result<i32, UblkError> {
-        self.get_inner_mut().get_queue_affinity_async(q, bm).await
-    }
 
     /// Set single CPU affinity for a specific queue
     ///
@@ -2467,26 +2309,6 @@ impl UblkCtrl {
         }
     }
 
-    /// Start user recover for this device asynchronously
-    ///
-    pub async fn start_user_recover_async(&self) -> Result<i32, UblkError> {
-        let mut count = 0u32;
-        let unit = 100_u32;
-
-        loop {
-            let res = self.get_inner_mut().__start_user_recover_async().await;
-            if let Ok(r) = res {
-                if r == -libc::EBUSY {
-                    futures_timer::Delay::new(std::time::Duration::from_millis(unit as u64)).await;
-                    count += unit;
-                    if count < 30000 {
-                        continue;
-                    }
-                }
-            }
-            return res;
-        }
-    }
 
     /// Start ublk device
     ///
@@ -2557,19 +2379,6 @@ impl UblkCtrl {
         ctrl.stop()
     }
 
-    /// Stop ublk device asynchronously
-    ///
-    /// Remove json export, and send stop command to control device asynchronously
-    ///
-    pub async fn stop_dev_async(&self) -> Result<i32, UblkError> {
-        let mut ctrl = self.get_inner_mut();
-        let rp = ctrl.run_path();
-
-        if ctrl.for_add_dev() && Path::new(&rp).exists() {
-            fs::remove_file(rp)?;
-        }
-        ctrl.stop_async().await
-    }
 
     /// Kill this device
     ///
@@ -2583,17 +2392,6 @@ impl UblkCtrl {
         self.get_inner_mut().stop()
     }
 
-    /// Kill this device asynchronously
-    ///
-    /// Preferred method for target code to stop & delete device,
-    /// which is safe and can avoid deadlock.
-    ///
-    /// But device may not be really removed yet, and the device ID
-    /// can still be in-use after kill_dev_async() returns.
-    ///
-    pub async fn kill_dev_async(&self) -> Result<i32, UblkError> {
-        self.get_inner_mut().stop_async().await
-    }
 
     /// Remove this device and its exported json file
     ///
@@ -2625,22 +2423,6 @@ impl UblkCtrl {
         Ok(0)
     }
 
-    /// Delete ublk device using async/await pattern
-    ///
-    /// This method provides true async/await support for device deletion,
-    /// using the async uring infrastructure for non-blocking operations.
-    /// This is an alternative to del_dev_async() that follows the established
-    /// async/await patterns used by other async methods in the API.
-    ///
-    pub async fn del_dev_async_await(&self) -> Result<i32, UblkError> {
-        let mut ctrl = self.get_inner_mut();
-
-        ctrl.del_async_await().await?;
-        if Path::new(&ctrl.run_path()).exists() {
-            fs::remove_file(ctrl.run_path())?;
-        }
-        Ok(0)
-    }
 
     /// Calculate queue affinity based on device settings
     ///
@@ -2658,21 +2440,6 @@ impl UblkCtrl {
             })
     }
 
-    async fn calculate_queue_affinity_async(&self, queue_id: u16) -> UblkQueueAffinity {
-        let affi = self
-            .get_inner_mut()
-            .create_thread_affinity_async(queue_id)
-            .await
-            .unwrap_or_else(|_| {
-                // Fallback to kernel affinity if thread affinity creation fails
-                let mut affinity = UblkQueueAffinity::new();
-                self.get_queue_affinity(queue_id as u32, &mut affinity)
-                    .unwrap_or_default();
-                affinity
-            });
-        log::info!("calculate queue affinity...done\n");
-        affi
-    }
 
     /// Set queue thread affinity using thread ID
     ///
@@ -2692,18 +2459,6 @@ impl UblkCtrl {
         }
     }
 
-    pub async fn set_thread_affinity_async(&self, qid: u16, tid: libc::pid_t) {
-        // Calculate and set affinity using the thread ID
-        let affinity = self.calculate_queue_affinity_async(qid).await;
-
-        unsafe {
-            libc::sched_setaffinity(
-                tid,
-                affinity.buf_len(),
-                affinity.addr() as *const libc::cpu_set_t,
-            );
-        }
-    }
 
     /// Initialize queue thread and return tid
     ///
@@ -2839,42 +2594,12 @@ impl UblkCtrl {
 mod tests {
     use crate::ctrl::{UblkCtrlBuilder, UblkQueueAffinity};
     use crate::io::{UblkDev, UblkIOCtx, UblkQueue};
-    use crate::uring_async::ublk_wake_task;
     use crate::UblkError;
     use crate::{ctrl::UblkCtrl, UblkFlags, UblkIORes};
     use std::cell::Cell;
     use std::path::Path;
     use std::rc::Rc;
 
-    /// Block on all tasks in the executor until they are finished
-    fn ublk_join_tasks<T>(
-        exe: &smol::LocalExecutor,
-        tasks: Vec<smol::Task<T>>,
-    ) -> Result<(), UblkError> {
-        use io_uring::{squeue, IoUring};
-        loop {
-            // Check if all tasks are finished
-            if tasks.iter().all(|task| task.is_finished()) {
-                break;
-            }
-
-            // Drive the executor to make progress on tasks
-            while exe.try_tick() {}
-
-            // Handle control uring events
-            let entry =
-                crate::ctrl::with_ctrl_ring_mut_internal!(|r: &mut IoUring<squeue::Entry128>| {
-                    match r.submit_and_wait(0) {
-                        Err(_) => None,
-                        _ => r.completion().next(),
-                    }
-                });
-            if let Some(cqe) = entry {
-                ublk_wake_task(cqe.user_data(), &cqe);
-            }
-        }
-        Ok(())
-    }
     #[test]
     fn test_ublk_get_features() {
         match UblkCtrl::get_features() {
@@ -2978,52 +2703,6 @@ mod tests {
         assert_eq!(inner.queue_selected_cpus[1], 0); // Should be initialized to 0
     }
 
-    #[test]
-    fn test_get_queue_affinity_async() {
-        let exe_rc = Rc::new(smol::LocalExecutor::new());
-        let exe = exe_rc.clone();
-
-        let job = exe_rc.spawn(async {
-            let ctrl = UblkCtrlBuilder::default()
-                .name("null_async_test")
-                .nr_queues(2_u16)
-                .dev_flags(UblkFlags::UBLK_DEV_F_ADD_DEV)
-                .build_async()
-                .await
-                .unwrap();
-
-            let mut affinity = UblkQueueAffinity::new();
-
-            // Test that method has correct signature and basic functionality
-            let result = ctrl.get_queue_affinity_async(0, &mut affinity).await;
-            match result {
-                Ok(_) => println!("✓ get_queue_affinity_async: Successfully retrieved affinity"),
-                Err(_) => println!(
-                    "✓ get_queue_affinity_async: Method exists and returns error as expected"
-                ),
-            }
-
-            // Verify it behaves consistently with the synchronous version for invalid queue
-            let mut sync_affinity = UblkQueueAffinity::new();
-            let mut async_affinity = UblkQueueAffinity::new();
-
-            let sync_result = ctrl.get_queue_affinity(999, &mut sync_affinity);
-            let async_result = ctrl
-                .get_queue_affinity_async(999, &mut async_affinity)
-                .await;
-
-            // Both should fail with the same type of error (though exact values may differ)
-            assert!(sync_result.is_err());
-            assert!(async_result.is_err());
-            let _ = ctrl.del_dev_async_await().await;
-        });
-
-        smol::block_on(exe_rc.run(async move {
-            let _ = ublk_join_tasks(&exe, vec![job]);
-        }));
-
-        println!("✓ get_queue_affinity_async method implemented correctly");
-    }
 
     #[test]
     fn test_get_queue_effective_affinity() {
@@ -3251,350 +2930,10 @@ mod tests {
         println!("  - Random CPU selection: PASS (selected CPU {})", cpu);
     }
 
-    /// Test async APIs
-    #[test]
-    fn test_async_apis() {
-        let _ = env_logger::builder()
-            .format_target(false)
-            .format_timestamp(None)
-            .try_init();
-        let exe_rc = Rc::new(smol::LocalExecutor::new());
-        let exe = exe_rc.clone();
 
-        log::info!("start async test");
-        let job = exe_rc.spawn(async {
-            log::info!("start main task");
-            // Test new_async with basic parameters
-            let result = UblkCtrlBuilder::default()
-                .name("test_async")
-                .dev_flags(UblkFlags::UBLK_DEV_F_ADD_DEV | UblkFlags::UBLK_CTRL_ASYNC_AWAIT)
-                .build_async()
-                .await;
 
-            // Should succeed or fail based on system capabilities, but method should exist
-            let ctrl = match result {
-                Ok(ctrl) => {
-                    let id = ctrl.dev_info().dev_id;
-                    println!("✓ new_async: Successfully created device {}", id);
-                    ctrl
-                }
-                Err(_e) => {
-                    println!("✓ new_async: Method exists and returns appropriate error");
-                    return;
-                }
-            };
 
-            if ctrl.read_dev_info_async().await.is_err() {
-                println!("✓ new_async: read_dev_info_async() failed");
-                return;
-            } else {
-                println!("✓ read_dev_info_async: Successfully read dev info")
-            }
 
-            let mut p = crate::sys::ublk_params {
-                ..Default::default()
-            };
 
-            if ctrl.get_params_async(&mut p).await.is_err() {
-                println!("✓ new_async: get_prarams_async() failed");
-            } else {
-                println!("✓ get_params_async: Successfully get parameters")
-            }
 
-            // Test get_queue_affinity_async
-            let mut affinity = UblkQueueAffinity::new();
-            match ctrl.get_queue_affinity_async(0, &mut affinity).await {
-                Ok(_) => {
-                    println!("✓ get_queue_affinity_async: Successfully retrieved queue affinity")
-                }
-                Err(_e) => println!(
-                    "✓ get_queue_affinity_async: Method exists and returns appropriate error"
-                ),
-            }
-
-            // Test dump_async method
-            match ctrl.dump_async().await {
-                Ok(()) => {
-                    println!("✓ dump_async: Successfully executed dump_async() method");
-                }
-                Err(e) => {
-                    println!(
-                        "✓ dump_async: Method exists and returns error as expected: {:?}",
-                        e
-                    );
-                }
-            }
-
-            if ctrl.stop_dev_async().await.is_err() {
-                println!("✓ new_async: stop_dev_async() failed");
-            } else {
-                println!("✓ stop_dev_async: Successfully")
-            }
-
-            if ctrl.del_dev_async_await().await.is_err() {
-                println!("✓ new_async: del_dev_async_await() failed");
-            } else {
-                println!("✓ del_dev_async_await: Successfully")
-            }
-
-            // Test new_simple_async
-            let result_simple = UblkCtrl::new_simple_async(ctrl.dev_info().dev_id as i32).await;
-            match result_simple {
-                Ok(_ctrl) => println!("✓ new_simple_async: Successfully created simple device"),
-                Err(_e) => {
-                    println!("✓ new_simple_async: Method exists and returns appropriate error")
-                }
-            }
-        });
-
-        smol::block_on(exe_rc.run(async move {
-            let _ = ublk_join_tasks(&exe, vec![job]);
-        }));
-
-        println!("✓ Async constructor methods are properly defined");
-    }
-
-    async fn io_async_fn(tag: u16, q: &UblkQueue<'_>) -> Result<(), UblkError> {
-        use crate::helpers::IoBuf;
-        use crate::BufDesc;
-
-        let mut res = 0;
-        let buf = IoBuf::<u8>::new(q.dev.dev_info.max_io_buf_bytes as usize);
-        let _buf = Some(buf);
-        let iod = q.get_iod(tag);
-        let buf_desc = BufDesc::Slice(_buf.as_ref().unwrap().as_slice());
-
-        // Submit initial prep command and handle any errors (including queue down)
-        // The IoBuf is automatically registered
-        q.submit_io_prep_cmd(tag, buf_desc.clone(), res, _buf.as_ref()).await?;
-
-        loop {
-            res = (iod.nr_sectors << 9) as i32;
-            // Any error (including QueueIsDown) will break the loop
-            q.submit_io_commit_cmd(tag, buf_desc.clone(), res).await?;
-        }
-    }
-
-    fn q_async_fn<'a>(
-        exe: &smol::LocalExecutor<'a>,
-        q_rc: &Rc<UblkQueue<'a>>,
-        depth: u16,
-        f_vec: &mut Vec<smol::Task<()>>,
-    ) {
-        for tag in 0..depth as u16 {
-            let q = q_rc.clone();
-            f_vec.push(exe.spawn(async move {
-                if let Err(e) = io_async_fn(tag, &q).await {
-                    match e {
-                        UblkError::QueueIsDown => {
-                            // Queue down is expected during shutdown, don't log as error
-                        }
-                        _ => {
-                            log::debug!("io_async_fn failed for tag {}: {}", tag, e);
-                        }
-                    }
-                }
-            }));
-        }
-    }
-
-    async fn device_handler_async(mlock_fail: bool) -> Result<(), UblkError> {
-        let ctrl = UblkCtrlBuilder::default()
-            .name("test_async")
-            .dev_flags(
-                UblkFlags::UBLK_DEV_F_ADD_DEV
-                    | UblkFlags::UBLK_CTRL_ASYNC_AWAIT
-                    | if mlock_fail {
-                        UblkFlags::UBLK_DEV_F_MLOCK_IO_BUFFER
-                    } else {
-                        UblkFlags::empty()
-                    },
-            )
-            .depth(8)
-            .build_async()
-            .await
-            .unwrap();
-
-        let tgt_init = |dev: &mut UblkDev| {
-            dev.set_default_params(250_u64 << 30);
-            Ok(())
-        };
-        let dev_arc = &std::sync::Arc::new(UblkDev::new(ctrl.get_name(), tgt_init, &ctrl)?);
-        let dev = dev_arc.clone();
-        assert!(dev_arc.dev_info.nr_hw_queues == 1);
-
-        // Todo: support to handle multiple queues in one thread context
-        let qh = std::thread::spawn(move || {
-            let q_rc = Rc::new(UblkQueue::new(0 as u16, &dev).unwrap());
-            let q = q_rc.clone();
-            let exe = smol::LocalExecutor::new();
-            let mut f_vec: Vec<smol::Task<()>> = Vec::new();
-
-            if mlock_fail {
-                q.mark_mlock_failed();
-            }
-
-            q_async_fn(&exe, &q, dev.dev_info.queue_depth as u16, &mut f_vec);
-
-            crate::uring_async::ublk_wait_and_handle_ios(&exe, &q_rc);
-            smol::block_on(exe.run(async { futures::future::join_all(f_vec).await }));
-        });
-
-        // Avoid to leak device
-        if let Err(_) = ctrl.start_dev_async(dev_arc).await {
-            log::warn!("device_handler_async: fail to start device(async)");
-        }
-
-        ctrl.dump_async().await?;
-        ctrl.kill_dev_async().await?;
-
-        // async/await needs to delete device by itself, otherwise we
-        // may hang in Drop() of UblkCtrlInner.
-        ctrl.del_dev_async_await().await?;
-
-        qh.join().unwrap_or_else(|_| {
-            eprintln!("dev-{} join queue thread failed", dev_arc.dev_info.dev_id)
-        });
-        Ok(())
-    }
-
-    /// Test async APIs for building ublk device
-    #[test]
-    fn test_create_ublk_async() {
-        let _ = env_logger::builder()
-            .format_target(false)
-            .format_timestamp(None)
-            .try_init();
-        let exe_rc = Rc::new(smol::LocalExecutor::new());
-        let exe = exe_rc.clone();
-        let mut fvec = Vec::new();
-
-        //support 64 devices
-        crate::ctrl::init_ctrl_task_ring_default(64 * 2).unwrap();
-
-        for _ in 0..64 {
-            fvec.push(exe_rc.spawn(async {
-                device_handler_async(false).await.unwrap();
-            }));
-        }
-
-        smol::block_on(exe_rc.run(async move {
-            let _ = ublk_join_tasks(&exe, fvec);
-        }));
-    }
-
-    #[test]
-    fn test_test_mlock_failure() {
-        let _ = env_logger::builder()
-            .format_target(false)
-            .format_timestamp(None)
-            .try_init();
-        let exe_rc = Rc::new(smol::LocalExecutor::new());
-        let exe = exe_rc.clone();
-
-        let io_task = exe_rc.spawn(async {
-            device_handler_async(true).await.unwrap();
-        });
-
-        smol::block_on(exe_rc.run(async move {
-            let _ = ublk_join_tasks(&exe, vec![io_task]);
-        }));
-    }
-
-    /// Test UBLK_CTRL_ASYNC_AWAIT flag enforcement
-    #[test]
-    fn test_ctrl_async_await_flag_enforcement() {
-        // Test with async flag support using a sync runtime context
-
-        let exe_rc = std::rc::Rc::new(smol::LocalExecutor::new());
-        let exe = exe_rc.clone();
-
-        let job = exe_rc.spawn(async move {
-            // Test with UBLK_CTRL_ASYNC_AWAIT flag set - should allow async but reject sync
-            let ctrl_async = UblkCtrlBuilder::default()
-                .name("test_async_flag")
-                .dev_flags(UblkFlags::UBLK_DEV_F_ADD_DEV | UblkFlags::UBLK_CTRL_ASYNC_AWAIT)
-                .build_async()
-                .await
-                .unwrap();
-
-            // Test with flag NOT set - should allow sync but reject async
-            let ctrl_sync = UblkCtrlBuilder::default()
-                .name("test_sync_flag")
-                .dev_flags(UblkFlags::UBLK_DEV_F_ADD_DEV)
-                .build()
-                .unwrap();
-
-            // Verify flags are set correctly
-            assert!(ctrl_async
-                .get_dev_flags()
-                .contains(UblkFlags::UBLK_CTRL_ASYNC_AWAIT));
-            assert!(!ctrl_sync
-                .get_dev_flags()
-                .contains(UblkFlags::UBLK_CTRL_ASYNC_AWAIT));
-
-            // Test sync API that should be rejected when UBLK_CTRL_ASYNC_AWAIT is set
-            {
-                let mut params = crate::sys::ublk_params {
-                    ..Default::default()
-                };
-                let sync_result = ctrl_async.get_params(&mut params);
-
-                match sync_result {
-                    Err(UblkError::OtherError(err)) => {
-                        assert_eq!(
-                            err,
-                            -libc::EPERM,
-                            "Sync API should be rejected with EPERM when UBLK_CTRL_ASYNC_AWAIT is set"
-                        );
-                    }
-                    _ => panic!("Sync API should fail when UBLK_CTRL_ASYNC_AWAIT flag is set"),
-                }
-            }
-
-            // Test async API that should be rejected when UBLK_CTRL_ASYNC_AWAIT is NOT set
-            {
-                let mut params = crate::sys::ublk_params {
-                    ..Default::default()
-                };
-                let async_result = ctrl_sync.get_params_async(&mut params).await;
-
-                match async_result {
-                    Err(UblkError::OtherError(err)) => {
-                        assert_eq!(err, -libc::EPERM, "Async API should be rejected with EPERM when UBLK_CTRL_ASYNC_AWAIT is not set");
-                    }
-                    _ => panic!("Async API should fail when UBLK_CTRL_ASYNC_AWAIT flag is not set"),
-                }
-            }
-
-            // Test async API that should work when UBLK_CTRL_ASYNC_AWAIT is set
-            {
-                let mut params = crate::sys::ublk_params {
-                    ..Default::default()
-                };
-                let async_result = ctrl_async.get_params_async(&mut params).await;
-
-                // The result may succeed or fail depending on system support,
-                // but it should NOT fail with EPERM (permission denied)
-                match async_result {
-                    Err(UblkError::OtherError(err)) => {
-                        assert_ne!(err, -libc::EPERM, "Async API should not be rejected with EPERM when UBLK_CTRL_ASYNC_AWAIT is set");
-                    }
-                    _ => {
-                        // Success or other errors are acceptable - we just care that EPERM is not returned
-                    }
-                }
-            }
-            let _ = ctrl_async.del_dev_async_await().await;
-        });
-
-        smol::block_on(exe_rc.run(async move {
-            let _ = ublk_join_tasks(&exe, vec![job]);
-        }));
-
-        println!("✓ UBLK_CTRL_ASYNC_AWAIT flag enforcement tests passed");
-        println!("  - Sync API rejection when flag is set: PASS");
-        println!("  - Async API rejection when flag is not set: PASS");
-        println!("  - Async API acceptance when flag is set: PASS");
-    }
 }
